Switched input size arrays to long int for consistency.

+++ src/LightField/include/LightField/LightField.h
@@ -15,7 +15,7 @@ public:
     LightField() = default;
     LightField(std::string rooth_path,std::string pattern);
     LightField(int numberOfCacheVerticalViews, int numberOfCacheHorizontalViews, int numberOfViewCacheLines);
-    LightField(std::array<int,5> sizes);
+    LightField(std::array<int64_t,5> sizes);
     int preSlantTan = 0;
     char *mViewFileNamePrefix;          /*!< lightfield view name: <mViewFileNamePrefix>_<horizontal index>_<vertical index>_<mViewFilenameSuffix>.pgm */
     char *mViewFileNameSuffix;          /*!< lightfield view name: <mViewFileNamePrefix>_<horizontal index>_<vertical index>_<mViewFilenameSuffix>.pgm */
@@ -49,7 +49,7 @@ public:
     void ReadBlock4DfromLightField(Block4D *targetBlock, int position_t, int position_s, int position_v, int position_u, int component=0);
     Block4D_ ReadBlock4DfromLightField_(std::array<int64_t,4> size, std::array<int64_t,4> position_t,int64_t channel );
     void WriteBlock4DtoLightField(Block4D *targetBlock, int position_t, int position_s, int position_v, int position_u, int component=0);
-    void WriteBlock4DtoLightField_(Block4D_ sourcelock, std::array<int,5> position);
+    void WriteBlock4DtoLightField_(Block4D_ sourcelock, std::array<int64_t,5> position);
     int FindViewFileName(char *viewFileName, int index_t, int index_s);
     void SetViewVerbosity(char verbosity);
